Can't test if running as root

+++ build.sh
@@ -10,7 +10,9 @@ docker build --force-rm --build-arg http_proxy="${http_proxy}" --build-arg https
 
 # Test that it can be ran
 docker run --rm "${IMAGE_NAME}:latest" /bin/env
-docker run --rm -v $(pwd):/home/build -w /home/build "${IMAGE_NAME}:latest" /bin/bash -c 'ls -l | grep "code_executor code_executor" || exit 1' || echo "User Mapping Failure" && exit 1
+if [[ $EUID -ne 0 ]]; then
+    docker run --rm -v $(pwd):/home/build -w /home/build "${IMAGE_NAME}:latest" /bin/bash -c 'ls -l | grep "code_executor code_executor" || exit 1' || echo "User Mapping Failure" && exit 1
+fi
 
 # Push to registry if configured
 if [ ! -z "${DOCKER_REGISTRY}" ]; then
